Replace rake.py contents with rake2.py contents
