+++ rake_spacy/__init__.py
@@ -6,5 +6,4 @@ from . import rake  # noqa: F401
 from . import scorers  # noqa: F401
 from . import stop_tokens  # noqa: F401
 from .__version__ import __version__  # noqa: F401
-from .rake import Metric  # noqa: F401
 from .rake import Rake  # noqa: F401
